Add to fixture instead of removing
