+++ subprojects/platform-native/src/testFixtures/groovy/org/gradle/nativeplatform/fixtures/app/SwiftAppWithLibraryTest.groovy
@@ -20,7 +20,7 @@ package org.gradle.nativeplatform.fixtures.app
  * A Swift app composed of 2 modules: an executable and a library.
  */
 class SwiftAppWithLibraryTest implements AppElement {
-    final library = new SwiftLibWithXCTest()
+    final library = new SwiftLibWithXCTestWithInfoPlist()
     final executable = new SwiftAppWithDep(library.main, library.main)
 
     @Override
