+++ subprojects/platform-native/src/testFixtures/groovy/org/gradle/nativeplatform/fixtures/app/SwiftLibTest.groovy
@@ -18,13 +18,23 @@ package org.gradle.nativeplatform.fixtures.app
 
 import org.gradle.integtests.fixtures.SourceFile
 
-class SwiftLibTest extends SwiftLibTestWithoutInfoPlist {
+class SwiftLibTest extends XCTestSourceElement {
+    final XCTestSourceFileElement sumTest
+    final XCTestSourceFileElement greeterTest
+    final XCTestSourceFileElement multiplyTest
+
     SwiftLibTest(GreeterElement greeter, SumElement sum, MultiplyElement multiply) {
-        super(greeter, sum, multiply)
+        sumTest = new SwiftSumTest("GreeterTest", sum).withImport("Greeter")
+        greeterTest = new SwiftGreeterTest("GreeterTest", greeter).withImport("Greeter")
+        multiplyTest = new SwiftMultiplyTest("GreeterTest", multiply).withTestableImport("Greeter")
+    }
+
+    List<XCTestSourceFileElement> getTestSuites() {
+        return [sumTest, greeterTest, multiplyTest]
     }
 
     @Override
     List<SourceFile> getFiles() {
-        return super.files + [sumTest.emptyInfoPlist()]
+        return super.files
     }
 }
