+++ subprojects/platform-native/src/testFixtures/groovy/org/gradle/nativeplatform/fixtures/app/SwiftXCTestBundleWithInfoPlist.groovy
@@ -18,23 +18,9 @@ package org.gradle.nativeplatform.fixtures.app
 
 import org.gradle.integtests.fixtures.SourceFile
 
-class SwiftLibTestWithoutInfoPlist extends XCTestSourceElement {
-    final XCTestSourceFileElement sumTest
-    final XCTestSourceFileElement greeterTest
-    final XCTestSourceFileElement multiplyTest
-
-    SwiftLibTestWithoutInfoPlist(GreeterElement greeter, SumElement sum, MultiplyElement multiply) {
-        sumTest = new SwiftSumTest("GreeterTest", sum).withImport("Greeter")
-        greeterTest = new SwiftGreeterTest("GreeterTest", greeter).withImport("Greeter")
-        multiplyTest = new SwiftMultiplyTest("GreeterTest", multiply).withTestableImport("Greeter")
-    }
-
-    List<XCTestSourceFileElement> getTestSuites() {
-        return [sumTest, greeterTest, multiplyTest]
-    }
-
+class SwiftXCTestBundleWithInfoPlist extends SwiftXCTestBundle {
     @Override
     List<SourceFile> getFiles() {
-        return super.files
+        super.files + [emptyInfoPlist()]
     }
 }
